+++ launcher-core/src/test/java/com/launcher/core/architecture/engine/LauncherEngineTest.java
@@ -4,6 +4,8 @@ import com.launcher.core.LauncherEngine;
 import com.launcher.core.architecture.support.RecordingManifestService;
 import com.launcher.core.architecture.support.RecordingOperationManager;
 import com.launcher.core.configuration.LauncherConfiguration;
+import com.launcher.core.download.DownloadPlan;
+import com.launcher.core.download.DownloadPlanBuilder;
 import com.launcher.core.event.EventBus;
 import com.launcher.core.manifest.ManifestService;
 import com.launcher.core.operation.result.OperationResult;
@@ -20,10 +22,10 @@ import java.net.URI;
 import java.nio.file.Path;
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.*;
 
 class LauncherEngineTest {
-
+    private final DownloadPlanBuilder builder = new DownloadPlanBuilder();
     private LauncherConfiguration configuration() {
         return new LauncherConfiguration(
                 URI.create("currentPath"),
@@ -31,35 +33,18 @@ class LauncherEngineTest {
         );
     }
 
-    private VerificationPlan getVerificationPlan(boolean isValid) {
+    private VerificationPlan getVerificationPlan(String path, VerificationStatus status) {
 
         return new VerificationPlan(
                 List.of(
                         new FileVerificationResult(
                                 new FileEntry(
-                                        "test-path",
-                                        "sha256",
+                                        path,
+                                        "sha256-" + path,
                                         321L,
-                                        "https://test-url.com"
+                                        "https://test-url.com/" + path
                                 ),
-                                isValid ? VerificationStatus.VALID : VerificationStatus.CORRUPTED
-                        )
-                )
-        );
-    }
-
-    private VerificationPlan getMissingVerificationPlan() {
-
-        return new VerificationPlan(
-                List.of(
-                        new FileVerificationResult(
-                                new FileEntry(
-                                        "missing-path",
-                                        "sha256222",
-                                        123,
-                                        "https://test-missing-url.com"
-                                ),
-                                VerificationStatus.MISSING
+                                status
                         )
                 )
         );
@@ -70,11 +55,131 @@ class LauncherEngineTest {
     }
 
     @Test
+    void should_transition_to_failed_when_build_download_plan_failed() {
+        //given
+        LauncherStateMachine stateMachine = stateMachine();
+        ManifestService manifestService = new RecordingManifestService();
+        VerificationPlan notValidVerificationPlan =
+                getVerificationPlan("not-valid.jar", VerificationStatus.MISSING);
+        RecordingOperationManager operationManager =
+                new RecordingOperationManager();
+
+        operationManager.registerResult(
+                OperationType.LOAD_MANIFEST,
+                OperationResult.success()
+        );
+
+        operationManager.registerBehavior(
+                OperationType.LOAD_MANIFEST,
+                context -> context.setManifest(manifestService.loadManifest())
+        );
+
+        operationManager.registerResult(
+                OperationType.VERIFY_FILES,
+                OperationResult.success()
+        );
+
+        operationManager.registerBehavior(
+                OperationType.VERIFY_FILES,
+                context -> context.setVerificationPlan(notValidVerificationPlan)
+        );
+
+        operationManager.registerResult(
+                OperationType.BUILD_DOWNLOAD_PLAN,
+                OperationResult.failure("failed to build download plan")
+        );
+
+        LauncherEngine engine = new LauncherEngine(stateMachine, operationManager);
+
+        //when
+        engine.launch(configuration());
+
+        //then
+        assertEquals(List.of(
+                OperationType.LOAD_MANIFEST,
+                OperationType.VERIFY_FILES,
+                OperationType.BUILD_DOWNLOAD_PLAN
+        ), operationManager.getExecutedOperationTypes());
+
+        assertEquals(
+                LauncherState.FAILED,
+                stateMachine.getCurrentState()
+        );
+
+        assertNull(operationManager.getReceivedContext().getDownloadPlan());
+    }
+
+    @Test
+    void should_execute_build_download_plan_when_verification_plan_is_not_valid() {
+        //given
+        LauncherStateMachine stateMachine = stateMachine();
+        ManifestService manifestService = new RecordingManifestService();
+        VerificationPlan notValidVerificationPlan =
+                getVerificationPlan("not-valid.jar", VerificationStatus.MISSING);
+        RecordingOperationManager operationManager =
+                new RecordingOperationManager();
+
+        operationManager.registerResult(
+                OperationType.LOAD_MANIFEST,
+                OperationResult.success()
+        );
+
+        operationManager.registerBehavior(
+                OperationType.LOAD_MANIFEST,
+                context -> context.setManifest(manifestService.loadManifest())
+        );
+
+        operationManager.registerResult(
+                OperationType.VERIFY_FILES,
+                OperationResult.success()
+        );
+
+        operationManager.registerBehavior(
+                OperationType.VERIFY_FILES,
+                context -> context.setVerificationPlan(notValidVerificationPlan)
+        );
+
+        operationManager.registerResult(
+                OperationType.BUILD_DOWNLOAD_PLAN,
+                OperationResult.success()
+        );
+
+        DownloadPlan plan = builder.build(notValidVerificationPlan);
+
+        operationManager.registerBehavior(
+                OperationType.BUILD_DOWNLOAD_PLAN,
+                context -> context.setDownloadPlan(plan)
+        );
+
+        LauncherEngine engine = new LauncherEngine(stateMachine, operationManager);
+
+        //when
+        engine.launch(configuration());
+
+        //then
+        assertEquals(List.of(
+                OperationType.LOAD_MANIFEST,
+                OperationType.VERIFY_FILES,
+                OperationType.BUILD_DOWNLOAD_PLAN
+        ), operationManager.getExecutedOperationTypes());
+
+        // Temporary behavior until the download flow is added
+        assertEquals(
+                LauncherState.FAILED,
+                stateMachine.getCurrentState()
+        );
+
+        assertEquals(plan, operationManager.getReceivedContext().getDownloadPlan());
+    }
+
+
+    @Test
     void should_execute_load_manifest_then_verify_files_when_launch_started() {
         //given
         LauncherStateMachine stateMachine = stateMachine();
         ManifestService manifestService = new RecordingManifestService();
-        VerificationPlan validVerificationPlan = getVerificationPlan(true);
+        VerificationPlan validVerificationPlan =
+                getVerificationPlan("valid.jar", VerificationStatus.VALID);
         RecordingOperationManager operationManager =
                 new RecordingOperationManager();
 
@@ -215,7 +320,7 @@ class LauncherEngineTest {
         //given
         LauncherStateMachine stateMachine = stateMachine();
         ManifestService manifestService = new RecordingManifestService();
-        VerificationPlan missingVerificationPlan = getMissingVerificationPlan();
+        VerificationPlan missingVerificationPlan = getVerificationPlan("missing.jar", VerificationStatus.MISSING);
         RecordingOperationManager operationManager =
                 new RecordingOperationManager();
 
@@ -247,7 +352,8 @@ class LauncherEngineTest {
         //then
         assertEquals(List.of(
                 OperationType.LOAD_MANIFEST,
-                OperationType.VERIFY_FILES
+                OperationType.VERIFY_FILES,
+                OperationType.BUILD_DOWNLOAD_PLAN
         ), operationManager.getExecutedOperationTypes());
 
         assertEquals(
@@ -261,7 +367,10 @@ class LauncherEngineTest {
         //given
         LauncherStateMachine stateMachine = stateMachine();
         ManifestService manifestService = new RecordingManifestService();
-        VerificationPlan invalidVerificationPlan = getVerificationPlan(false);
+        VerificationPlan invalidVerificationPlan = getVerificationPlan(
+                "not-valid.jar",
+                VerificationStatus.CORRUPTED
+        );
         RecordingOperationManager operationManager =
                 new RecordingOperationManager();
 
@@ -293,7 +402,8 @@ class LauncherEngineTest {
         //then
         assertEquals(List.of(
                 OperationType.LOAD_MANIFEST,
-                OperationType.VERIFY_FILES
+                OperationType.VERIFY_FILES,
+                OperationType.BUILD_DOWNLOAD_PLAN
         ), operationManager.getExecutedOperationTypes());
 
         assertEquals(
@@ -303,10 +413,58 @@ class LauncherEngineTest {
     }
 
     @Test
+    void should_not_build_download_plan_when_verification_plan_is_valid() {
+        //given
+        LauncherStateMachine stateMachine = stateMachine();
+        VerificationPlan validVerificationPlan = getVerificationPlan("valid.jar", VerificationStatus.VALID);
+        ManifestService manifestService = new RecordingManifestService();
+        RecordingOperationManager operationManager =
+                new RecordingOperationManager();
+
+        operationManager.registerResult(
+                OperationType.LOAD_MANIFEST,
+                OperationResult.success()
+        );
+
+        operationManager.registerBehavior(
+                OperationType.LOAD_MANIFEST,
+                context -> context.setManifest(manifestService.loadManifest())
+        );
+
+        operationManager.registerResult(
+                OperationType.VERIFY_FILES,
+                OperationResult.success()
+        );
+
+        operationManager.registerBehavior(
+                OperationType.VERIFY_FILES,
+                context -> context.setVerificationPlan(validVerificationPlan)
+        );
+
+        LauncherEngine engine = new LauncherEngine(stateMachine, operationManager);
+
+        //when
+        engine.launch(configuration());
+
+        //then
+        assertEquals(List.of(
+                OperationType.LOAD_MANIFEST,
+                OperationType.VERIFY_FILES
+        ), operationManager.getExecutedOperationTypes());
+
+        assertEquals(
+                LauncherState.RUNNING,
+                stateMachine.getCurrentState()
+        );
+
+        assertNull(operationManager.getReceivedContext().getDownloadPlan());
+    }
+
+    @Test
     void should_transition_to_running_when_manifest_loaded_and_files_are_valid() {
         //given
         LauncherStateMachine stateMachine = stateMachine();
-        VerificationPlan validVerificationPlan = getVerificationPlan(true);
+        VerificationPlan validVerificationPlan = getVerificationPlan("valid.jar", VerificationStatus.VALID);
         ManifestService manifestService = new RecordingManifestService();
         RecordingOperationManager operationManager =
                 new RecordingOperationManager();